+++ test/helpers/maru_response_test.exs
@@ -1,35 +1,7 @@
 defmodule Maru.Helpers.ResponseTest do
   use ExUnit.Case, async: true
   import Plug.Test
-  import Plug.Conn
   use Maru.Helpers.Response
-
-  test "json response" do
-    conn = conn(:get, "/")
-    conn = %{hello: :world} |> json
-    assert conn.state == :sent
-    assert conn.status == 200
-    assert get_resp_header(conn, "content-type") == ["application/json; charset=utf-8"]
-    assert conn.resp_body == ~s({"hello":"world"})
-  end
-
-  test "html response" do
-    conn = conn(:get, "/")
-    conn = "<html>hello world<html>" |> html
-    assert conn.state == :sent
-    assert conn.status == 200
-    assert get_resp_header(conn, "content-type") == ["text/html; charset=utf-8"]
-    assert conn.resp_body == "<html>hello world<html>"
-  end
-
-  test "text response" do
-    conn = conn(:get, "/")
-    conn = "hello world" |> text
-    assert conn.state == :sent
-    assert conn.status == 200
-    assert get_resp_header(conn, "content-type") == ["text/plain; charset=utf-8"]
-    assert conn.resp_body == "hello world"
-  end
 
   test "header" do
     conn = conn(:get, "/") |> put_req_header("foo", "bar")
@@ -42,5 +14,17 @@ defmodule Maru.Helpers.ResponseTest do
     conn = conn(:get, "/")
     assign(:foo, "bar")
     assert assigns == %{foo: "bar"}
+  end
+
+  test "status int" do
+    conn = conn(:post, "/")
+    status 201
+    assert 201 = conn.status
+  end
+
+  test "status atom" do
+    conn = conn(:get, "/")
+    status :ok
+    assert 200 = conn.status
   end
 end